Divide files into components
