--- src/App.css
@@ -1,280 +1,0 @@
-/* Table of Content
-==================================================
-SECTION:
-	header
-	product
-	cart
-*/
-
-/* BASE - Base tyles, Variables, Mixins, etc. */
-html {
-	width: 1080px;
-	height: 1920px;
-	overflow-x: scroll;
-}
-
-body {
-  margin: 0px;
-  background-color: #ffffff;
-  font-family: "Lato", sans-serif;
-  color: #333333;
-  font-size: 16px;
-}
-ul {
-  float: left;
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-a {
-  text-decoration: none;
-  color: #333333;
-}
-a:hover,
-a:focus {
-  outline: none;
-  text-decoration: none;
-}
-
-h1, h2, h3, h4, h5, h6 {
-  font-family: "Raleway", sans-serif;
-}
-
-h2 {
-  font-size: 30px;
-  font-weight: 700;
-  line-height: 40px;
-  margin: 0;
-  padding-bottom: 10px;
-}
-img {
-  border: none;
-}
-button {
-	border: 0px;
-}
-
-
-
-/*
-
-Top(Header) Part
-
-*/
-
-#header {
-	background-color: #519d9e;
-	width: 1080px;
-	height: 360px;
-}
-#header .title {
-	width: 1080px;
-	height: 200px;
-	position: relative;
-}
-#header .title .big-title {
-	position: relative;
-	top: 50px;
-	left: 70px;
-	height: 54px;
-	color: #ffffff;
-	font-size: 48px;
-	font-weight: 700;
-	letter-spacing: -.483333px;
-}
-#header .title .small-title {
-	position: relative;
-	top: 54px;
-	left: 70px;
-	height: 38px;
-	color: #ffffff;
-	font-size: 34px;
-	font-weight: 500;
-	letter-spacing: -.483333px;
-}
-#header .category {
-	background-color: #e4e7e8;
-	width: 1080px;
-	height: 80px;
-}
-#header .category .category-button {
-	padding-top: 24px;
-	display: inline-block;
-	margin-left: 85px;
-	margin-right: 85px;
-
-	font-size: 30px;
-	font-weight: 500;
-}
-#header .category .category-button a {
-	color: #b1b1b1;
-}
-#header .find {
-	background-color: #f0f3f4;
-	width: 1080px;
-	height: 80px;
-}
-#header .find .search-box {
-	border: 1px solid #ddd;
-	border-radius: 8px;
-	background-color: #ffffff;
-	width: 700px;
-	height: 56px;
-	position: relative;
-	top: 10px;
-	left: 190px;
-
-	font-family: Noto Sans CJK KR;
-	font-size: 24px;
-	font-weight: 500;
-	color: #555;
-	text-align: center;
-}
-#header .find .search-button {
-	border: solid #b1b1b1;
-	border-width: 2px;
-	border-radius: 4px;
-	background-color: transparent;
-	height: 50px;
-	width: 80px;
-	position: relative;
-	top: 18px;
-	left: 100px;
-}
-#header .find .search-button img {
-	width: 32px;
-	height: 32px;
-}
-
-/*
-
-Middle(Product) Part
-
-*/
-
-#product {
-	width: 1080px;
-	height: 1150px;
-	background-color: #f0f3f4;
-}
-#product .menu {
-	width: 1080px;
-	height: 60px;
-	background-color: #9dc7c8;
-	overflow: auto;
-}
-#product .menu .menu-button {
-	padding-top: 12px;
-	display: inline-block;
-	margin-left: 35px;
-	margin-right: 35px;
-	font-size: 26px;
-	font-weight: 500;
-}
-#product .menu .menu-button li, a {
-	overflow: hidden;
-	white-space: nowrap;
-}
-#product .menu a {
-	color: #f2efef;
-	overflow: hidden;
-	white-space: nowrap;
-}
-#product .main-commerce {
-	width: 1080px;
-	height: 1090px;
-}
-#product .container {
-	margin-top: 55px;
-	margin-left: 50px;
-	margin-right: 50px;
-}
-#product .container .item {
-	display: inline;
-	float: left;
-	width: 300px;
-	height: 320px;
-	background-color: #ffffff;
-	border-radius: 10px;
-	box-shadow: 0 3px 6px rgba(0, 0, 0, .16);
-	margin: 10px;
-}
-#product .container .item .item-image {
-	height: 240px;
-}
-#product .container .item .description {
-	height: 80px;
-}
-#product .container .item .description .name {
-	font-size: 16px;
-	font-weight: 700;
-	color: #303336;
-	position: relative;
-	text-align: right;
-	right: 10px;
-}
-#product .container .item .description .price {
-	font-weight: 500;
-	color: #7b8591;
-	text-align: right;
-	position: relative;
-	right: 18px;
-}
-
-/*
- 
-Bottom(Cart) Part
- 
-*/
-
-
-#cart {
-	width: 1080px;
-	height: 410px;
-	bottom: 0px;
-	font-size: 26px;
-}
-#cart .cart-top {
-	height: 66px;
-	background-color: #9dc7c8;
-}
-#cart .cart-top .cancel-button {
-	border: 0px;
-	background-color: #f0f3f4;
-	border-radius: 10px;
-	box-shadow: 3px 3px 3px rgba(0, 0, 0, .16);
-	width: 180px;
-	height: 50px;
-
-	text-align: center;
-	font-size: 26px;
-	font-weight: 500;
-	color: #000000;
-
-	position: relative;
-	top: 8px;
-	left: 840px;
-}
-#cart .body {
-	background-color: #e4e7e8;
-	width: 1080px;
-	height: 344px;
-}
-#cart .body .purchase-button {
-	border: 0px;
-	background-color: #519d9e;
-	border-radius: 10px;
-	box-shadow: 4px 4px 4px rgba(0, 0, 0, .16);
-	width: 180px;
-	height: 275px;
-
-	text-align: center;
-	font-size: 28px;
-	font-weight: 500;
-	color: #ffffff;
-
-	position: relative;
-	top: 40px;
-	left: 840px;
-}
